x86/cpu: Introduce INTEL_CPU_FAM*() helper macros
These macros are often used by drivers and we have already a lot of
duplication as ICPU() macro across the drivers.

Provide a generic x86 macro for users.

Note, as Ingo Molnar pointed out this has a hidden issue when a driver
needs to preserve const qualifier. Though, it would be addressed
separately at some point.

[1]: https://lkml.org/lkml/2017/12/28/85

Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>